+++ src/main/java/insane96mcp/iguanatweaksexpanded/module/experience/enchantments/enchantment/curse/CurseOfSteelFall.java
@@ -1,0 +1,48 @@
+package insane96mcp.iguanatweaksexpanded.module.experience.enchantments.enchantment.curse;
+
+import insane96mcp.iguanatweaksexpanded.module.experience.enchantments.NewEnchantmentsFeature;
+import insane96mcp.insanelib.world.enchantments.IEnchantmentTooltip;
+import net.minecraft.ChatFormatting;
+import net.minecraft.network.chat.Component;
+import net.minecraft.world.entity.EquipmentSlot;
+import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.item.enchantment.Enchantment;
+import net.minecraft.world.item.enchantment.EnchantmentCategory;
+import net.minecraft.world.item.enchantment.EnchantmentHelper;
+import net.minecraftforge.event.entity.living.LivingFallEvent;
+
+public class CurseOfSteelFall extends Enchantment implements IEnchantmentTooltip {
+    public CurseOfSteelFall() {
+        super(Rarity.RARE, EnchantmentCategory.ARMOR_FEET, EquipmentSlot.values());
+    }
+
+    @Override
+    public boolean isTreasureOnly() {
+        return true;
+    }
+
+    @Override
+    public boolean isCurse() {
+        return true;
+    }
+
+    public int getMinCost(int pEnchantmentLevel) {
+        return 25;
+    }
+
+    public int getMaxCost(int pEnchantmentLevel) {
+        return 50;
+    }
+
+    public static void onFall(LivingFallEvent event) {
+        int lvl = EnchantmentHelper.getEnchantmentLevel(NewEnchantmentsFeature.CURSE_OF_STEEL_FALL.get(), event.getEntity());
+        if (lvl <= 0)
+            return;
+        event.setDistance(event.getDistance() * 1.5f);
+    }
+
+    @Override
+    public Component getTooltip(ItemStack itemStack, int i) {
+        return Component.translatable(this.getDescriptionId() + ".tooltip").withStyle(ChatFormatting.RED);
+    }
+}